chore(mobile): split out Timeline dragging into new widget
